@@ -1,40 +1,39 @@
-'use strict';
-var fs = require('fs');
-var path = require('path');
-var assert = require('assert');
-var gutil = require('gulp-util');
-var unzip = require('decompress-unzip');
-var vinylAssign = require('vinyl-assign');
-var vinylFile = require('vinyl-file');
-var zip = require('./');
+import fs from 'fs';
+import path from 'path';
+import test from 'ava';
+import gutil from 'gulp-util';
+import unzip from 'decompress-unzip';
+import vinylAssign from 'vinyl-assign';
+import vinylFile from 'vinyl-file';
+import zip from '.';
 
-it('should zip files', function (cb) {
-	var stream = zip('test.zip');
-	var unzipper = unzip();
-	var stats = fs.statSync(__dirname + '/fixture/fixture.txt');
-	var files = [];
+test.cb('should zip files', t => {
+	const stream = zip('test.zip');
+	const unzipper = unzip();
+	const stats = fs.statSync(path.join(__dirname, 'fixture/fixture.txt'));
+	const files = [];
 
 	unzipper.on('data', files.push.bind(files));
-	unzipper.on('end', function () {
-		assert.equal(files[0].path, 'fixture.txt');
-		assert.equal(files[1].path, 'fixture2.txt');
-		assert.equal(files[0].contents.toString(), 'hello world');
-		assert.equal(files[1].contents.toString(), 'hello world 2');
-		assert.equal(files[0].stat.mode, stats.mode);
-		assert.equal(files[1].stat.mode, stats.mode);
-		cb();
+	unzipper.on('end', () => {
+		t.is(files[0].path, 'fixture.txt');
+		t.is(files[1].path, 'fixture2.txt');
+		t.is(files[0].contents.toString(), 'hello world');
+		t.is(files[1].contents.toString(), 'hello world 2');
+		t.is(files[0].stat.mode, stats.mode);
+		t.is(files[1].stat.mode, stats.mode);
+		t.end();
 	});
 
-	stream.on('data', function (file) {
-		assert.equal(path.normalize(file.path), path.join(__dirname, 'fixture/test.zip'));
-		assert.equal(file.relative, 'test.zip');
-		assert(file.contents.length > 0);
+	stream.on('data', file => {
+		t.is(path.normalize(file.path), path.join(__dirname, 'fixture/test.zip'));
+		t.is(file.relative, 'test.zip');
+		t.true(file.contents.length > 0);
 	});
 
 	stream.write(new gutil.File({
 		cwd: __dirname,
-		base: __dirname + '/fixture',
-		path: __dirname + '/fixture/fixture.txt',
+		base: path.join(__dirname, 'fixture'),
+		path: path.join(__dirname, 'fixture/fixture.txt'),
 		contents: new Buffer('hello world'),
 		stat: {
 			mode: stats.mode,
@@ -44,8 +43,8 @@ it('should zip files', function (cb) {
 
 	stream.write(new gutil.File({
 		cwd: __dirname,
-		base: __dirname + '/fixture',
-		path: __dirname + '/fixture/fixture2.txt',
+		base: path.join(__dirname, 'fixture'),
+		path: path.join(__dirname, 'fixture/fixture2.txt'),
 		contents: new Buffer('hello world 2'),
 		stat: {
 			mode: stats.mode,
@@ -53,65 +52,65 @@ it('should zip files', function (cb) {
 		}
 	}));
 
-	stream.pipe(vinylAssign({extract:true})).pipe(unzipper);
+	stream.pipe(vinylAssign({extract: true})).pipe(unzipper);
 	stream.end();
 });
 
-it('should zip files (using streams)', function (cb) {
-	var file = vinylFile.readSync(path.join(__dirname, 'fixture/fixture.txt'), {buffer: false});
-	var stats = fs.statSync(__dirname + '/fixture/fixture.txt');
-	var stream = zip('test.zip');
-	var unzipper = unzip();
-	var files = [];
+test.cb('should zip files (using streams)', t => {
+	const file = vinylFile.readSync(path.join(__dirname, 'fixture/fixture.txt'), {buffer: false});
+	const stats = fs.statSync(path.join(__dirname, 'fixture/fixture.txt'));
+	const stream = zip('test.zip');
+	const unzipper = unzip();
+	const files = [];
 
 	unzipper.on('data', files.push.bind(files));
-	unzipper.on('end', function () {
-		assert.equal(files[0].path, 'fixture/fixture.txt');
-		assert.equal(files[0].contents.toString(), 'hello world\n');
-		assert.equal(files[0].stat.mode, stats.mode);
-		cb();
+	unzipper.on('end', () => {
+		t.is(files[0].path, 'fixture/fixture.txt');
+		t.is(files[0].contents.toString(), 'hello world\n');
+		t.is(files[0].stat.mode, stats.mode);
+		t.end();
 	});
 
-	stream.on('data', function (file) {
-		assert.equal(path.normalize(file.path), path.join(__dirname, 'test.zip'));
-		assert.equal(file.relative, 'test.zip');
-		assert(file.contents.length > 0);
+	stream.on('data', file => {
+		t.is(path.normalize(file.path), path.join(__dirname, 'test.zip'));
+		t.is(file.relative, 'test.zip');
+		t.true(file.contents.length > 0);
 	});
 
-	stream.pipe(vinylAssign({extract:true})).pipe(unzipper);
+	stream.pipe(vinylAssign({extract: true})).pipe(unzipper);
 	stream.end(file);
 });
 
-it('should not skip empty directories', function (cb) {
-	var stream = zip('test.zip');
-	var unzipper = unzip();
-	var files = [];
-	var stats = {
-		isDirectory: function () {
+test.cb('should not skip empty directories', t => {
+	const stream = zip('test.zip');
+	const unzipper = unzip();
+	const files = [];
+	const stats = {
+		isDirectory() {
 			return true;
 		}
 	};
 
 	unzipper.on('data', files.push.bind(files));
-	unzipper.on('end', function () {
-		assert.equal(files[0].path, 'foo');
-		cb();
+	unzipper.on('end', () => {
+		t.is(files[0].path, 'foo');
+		t.end();
 	});
 
-	stream.on('data', function (file) {
-		assert.equal(path.normalize(file.path), path.join(__dirname, 'test.zip'));
-		assert.equal(file.relative, 'test.zip');
-		assert(file.contents.length > 0);
+	stream.on('data', file => {
+		t.is(path.normalize(file.path), path.join(__dirname, 'test.zip'));
+		t.is(file.relative, 'test.zip');
+		t.true(file.contents.length > 0);
 	});
 
 	stream.write(new gutil.File({
 		cwd: __dirname,
 		base: __dirname,
-		path: __dirname + '/foo',
+		path: path.join(__dirname, 'foo'),
 		contents: null,
 		stat: stats
 	}));
 
-	stream.pipe(vinylAssign({extract:true})).pipe(unzipper);
+	stream.pipe(vinylAssign({extract: true})).pipe(unzipper);
 	stream.end();
 });
